Check imports method
