+++ model/imports.py
@@ -1,0 +1,10 @@
+
+def check_imports():
+    result = True
+    try:
+        import shapely
+        # etc
+    except:
+        result = False
+
+    return result
